build: migrate to `tsdown`
