+++ package.json
@@ -43,8 +43,6 @@
     "@antfu/ni": "catalog:dev",
     "@arethetypeswrong/cli": "catalog:test",
     "@iconify/json": "catalog:dev",
-    "@rollup/plugin-json": "catalog:dev",
-    "@rollup/plugin-replace": "catalog:dev",
     "@shikijs/vitepress-twoslash": "catalog:docs",
     "@type-challenges/utils": "catalog:types",
     "@types/md5": "catalog:types",
@@ -92,15 +90,12 @@
     "prettier": "catalog:dev",
     "remove-markdown": "catalog:docs",
     "rolldown-vite": "catalog:dev",
-    "rollup": "catalog:dev",
-    "rollup-plugin-dts": "catalog:dev",
-    "rollup-plugin-esbuild": "catalog:dev",
-    "rollup-plugin-pure": "catalog:dev",
     "sharp": "catalog:dev",
     "simple-git": "catalog:dev",
     "simple-git-hooks": "catalog:dev",
     "taze": "catalog:dev",
     "tinyglobby": "catalog:dev",
+    "tsdown": "catalog:dev",
     "tsx": "catalog:dev",
     "typescript": "catalog:dev",
     "unocss": "catalog:dev",
@@ -126,7 +121,6 @@
     "@vueuse/rxjs": "workspace:*",
     "@vueuse/shared": "workspace:*",
     "eslint": "catalog:dev",
-    "rollup": "catalog:dev",
     "vite": "catalog:dev",
     "vite-plugin-pwa": "catalog:docs",
     "vitepress>vite": "npm:rolldown-vite@latest",
