+++ packages/integrations/package.json
@@ -40,18 +40,18 @@
     "./useSortable": "./useSortable.mjs",
     "./useSortable/component": "./useSortable/component.mjs"
   },
-  "main": "./index.mjs",
-  "module": "./index.mjs",
+  "main": "./index.js",
+  "module": "./index.js",
   "unpkg": "./index.iife.min.js",
   "jsdelivr": "./index.iife.min.js",
-  "types": "./index.d.mts",
+  "types": "./index.d.ts",
   "files": [
-    "**/*.d.mts",
+    "**/*.d.ts",
     "**/*.js",
     "**/*.mjs"
   ],
   "scripts": {
-    "build": "rollup --config=rollup.config.ts --configPlugin=rollup-plugin-esbuild",
+    "build": "tsdown",
     "prepack": "pnpm run build",
     "test:attw": "attw --pack --config-path ../../.attw.json ."
   },
