+++ packages/metadata/package.json
@@ -21,21 +21,21 @@
   ],
   "sideEffects": false,
   "exports": {
-    ".": "./index.mjs",
+    ".": "./index.js",
     "./*": "./*"
   },
-  "main": "./index.mjs",
-  "module": "./index.mjs",
-  "types": "./index.d.mts",
+  "main": "./index.js",
+  "module": "./index.js",
+  "types": "./index.d.ts",
   "files": [
-    "*.d.mts",
+    "*.d.ts",
     "*.js",
     "*.mjs",
     "index.json"
   ],
   "scripts": {
     "update": "tsx scripts/update.ts",
-    "build": "rollup --config=rollup.config.ts --configPlugin=rollup-plugin-esbuild",
+    "build": "tsdown",
     "prepack": "pnpm run build",
     "test:attw": "attw --pack --config-path ../../.attw.json ."
   }
